Added Java files to debug
I give up, I couldn't figure it out in C++. I wrote a small Java program
that we can use to debug. It seems like the depth of the solution
matters as well. If you change the max depth in Java to 1, it will block
the disjoint set (XX_X, it will go into the blank) but if you change the
max depth to 2, it will no longer block it.